requirements.txt
pytorch_lightning
sinabs
torchvision
mlflow